+++ bin/install-ubuntu-deps.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-sudo apt install librabbitmq-dev
+sudo apt install librabbitmq-dev libzmq3-dev