+++ CHANGELOG.md
@@ -9,3 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic extendable node architecture that builds on a four step system, pull items,
   transform items, place items, and handling failures based on previous steps.
 - Basic rabbit-mq node with a bear-bones implementation.
+- Started basic high level docs.
+- Added MMOP/0 worker-ready.
+- Added MMOP/0 start-node.
+- Added MMOP/0 start-node-success.
+- Added MMOP/0 start-node-failure.
+- Added node recipe system for constructing node threads.